+++ quac.h
@@ -17,7 +17,7 @@
 #include "item.h"
 
 #define QUAC_NAME    "Quac!"
-#define QUAC_VERSION "v0.8.0"
+#define QUAC_VERSION "v0.9.0"
 #define QUAC_ABOUT                                    \
     "Quick Action remote control\n" QUAC_VERSION "\n" \
     "github.com/rdefeo/quac"
@@ -58,6 +58,7 @@ typedef struct App {
         QuacAppLayout layout; // Defaults to Portrait
         bool show_icons; // Defaults to True
         bool show_headers; // Defaults to True
+        uint32_t subghz_duration; // Defaults to 1500 ms
         uint32_t rfid_duration; // Defaults to 2500 ms
         uint32_t nfc_duration; // Defaults to 1000 ms
         uint32_t ibutton_duration; // Defaults to 1000 ms