+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Auto-detect External SubGhz Antenna and use/prefer if available
 - Removed SugGhz External Antenna setting due to new auto-detect feature
 - Removed SubGhz Repeat setting, no longer needed (was non-standard)
+- Added SubGhz Duration setting for non-RAW signals
 
 ## 0.8